+++ src/collective/eeafaceted/collectionwidget/vocabulary.py
@@ -37,7 +37,7 @@ class CollectionVocabulary(object):
                             redirect_to = self._compute_redirect_to(collection, criterion)
                             break
 
-            items.append(SimpleTerm(brain.UID,
+            items.append(SimpleTerm(brain.getObject(),
                                     brain.UID,
                                     (safe_unicode(brain.Title),
                                      redirect_to)))
@@ -102,7 +102,7 @@ class CollectionCategoryVocabulary(object):
         while IFacetedNavigable.providedBy(root.aq_inner.aq_parent):
             root = root.aq_inner.aq_parent
         adapter = getAdapter(root, ICollectionCategories)
-        items = [SimpleTerm(key, key, value) for key, value in adapter.values]
+        items = [SimpleTerm(value, token, safe_unicode(value.Title())) for token, value in adapter.values]
         return SimpleVocabulary(items)